Set _array_indices to None (#18)
Before, this set _array_indices to [] if it was None/NULL before the migration. Now it carries over the None/NULL value. New entries to the database also use None/NULL.

A value of [] caused problems in pyuit/pbs_script.py:202 number_of_sub_jobs() "if self._array_indices is not None:" when displaying the jobs_table. That line is now fixed, but it also causes an error a few lines above during a job restart.