Fix process_results() not calling get_remote_files()
After a job completed, this no longer downloaded remote files because of
switching to the AsyncUitClient. This change uses async and await for
get_remote_files(). It used to wander into Tethys code with
process_results() and _process_results(), so I recreated that in
async_process_results(). I also had to await the async_process_results()
because it has the @database_sync_to_async decorator.

I don't use transfer_intermediate_files, so I have not been able to test
that change.